Support aggregation periods > 1 second
This patch allows carbonmem to store metrics in buckets larger than 1
second, for example 10s or 1 minute.

This means that t0 % bucketSize =  0, and we have to scale appropriately
when moving between times and array indices.

This also updates the 'seconds' parameter in a few places to be an
int32, to keep the same type as the rest of the time units, and to
remove conversions in a few expressions.