Fix trigram indexing bug
This fixes a bug in trigram indexing that was introduced in
ff44b284c8d6fde2edffedca3d6201cd7fd4dc9a .

Once we started using the freelist, document IDs were no longer being
inserted in sorted order, and the document IDs wouldn't match the path
numbers.  This would cause globbing queries to fail as the query code
assumes all the posting lists are sorted, and even if the IDs did come
back they might not back the entries in l.paths[].

This patch addresses this by only using the free list only if
we're not using the trigram index.